Add Spring Boot generator
