--- MassBank-Project/MassBank-OpenAPI/src/main/webapp/WEB-INF/web.xml
@@ -1,65 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<web-app xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xmlns="http://xmlns.jcp.org/xml/ns/javaee"
-	xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/javaee
-                      http://xmlns.jcp.org/xml/ns/javaee/web-app_4_0.xsd"
-	version="4.0">
-
-    <servlet>
-        <servlet-name>jersey</servlet-name>
-        <servlet-class>org.glassfish.jersey.servlet.ServletContainer</servlet-class>
-        <init-param>
-            <param-name>jersey.config.server.provider.packages</param-name>
-            <param-value>
-                io.swagger.jaxrs.listing,
-                io.swagger.sample.resource,
-                org.openapitools.api
-            </param-value>
-        </init-param>
-        <init-param>
-            <param-name>jersey.config.server.provider.classnames</param-name>
-            <param-value>org.glassfish.jersey.media.multipart.MultiPartFeature</param-value>
-        </init-param>
-        <init-param>
-            <param-name>jersey.config.server.wadl.disableWadl</param-name>
-            <param-value>true</param-value>
-        </init-param>
-        <load-on-startup>1</load-on-startup>
-    </servlet>
-
-    <servlet>
-        <servlet-name>Jersey2Config</servlet-name>
-        <servlet-class>io.swagger.jersey.config.JerseyJaxrsConfig</servlet-class>
-        <init-param>
-            <param-name>api.version</param-name>
-            <param-value>1.0.0</param-value>
-        </init-param>
-        <init-param>
-            <param-name>swagger.api.title</param-name>
-            <param-value>OpenAPI Server</param-value>
-        </init-param>
-        <init-param>
-            <param-name>swagger.api.basepath</param-name>
-            <param-value>https://msbi.ipb-halle.de/MassBank-OpenAPI</param-value>
-        </init-param>
-
-        <load-on-startup>2</load-on-startup>
-    </servlet>
-    <servlet>
-        <servlet-name>Bootstrap</servlet-name>
-        <servlet-class>org.openapitools.api.Bootstrap</servlet-class>
-        <load-on-startup>2</load-on-startup>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>jersey</servlet-name>
-        <url-pattern>/MassBank-OpenAPI/*</url-pattern>
-    </servlet-mapping>
-    <filter>
-        <filter-name>ApiOriginFilter</filter-name>
-        <filter-class>org.openapitools.api.ApiOriginFilter</filter-class>
-    </filter>
-    <filter-mapping>
-        <filter-name>ApiOriginFilter</filter-name>
-        <url-pattern>/*</url-pattern>
-    </filter-mapping>
-</web-app>
